Restructure modules folder and pages
