+++ src/modules/chat/AboutChatPage.tsx
@@ -1,0 +1,3 @@
+export default function AboutPage() {
+  return <div>AboutPage</div>
+}
