+++ src/modules/chat/ChatPage/index.tsx
@@ -1,3 +1,4 @@
 import ChatPage from './ChatPage'
 
+export * from './ChatPage'
 export default ChatPage
