+++ src/modules/chat/HomePage/index.tsx
@@ -1,3 +1,4 @@
 import HomePage from './HomePage'
 
+export * from './HomePage'
 export default HomePage
